[FEAT] : 카테고리 버튼 클릭 시 데이터 수정되도록 수정

+++ Money-Planner/Money-Planner/Common/Progress.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-func getProgress(numerator : Int, denominator : Int) -> CGFloat{
+func getProgress(numerator : Int64, denominator : Int64) -> CGFloat{
     let progress = (Double(numerator)/Double(denominator)) > 1.0 ? 1 : CGFloat(Double(numerator)/Double(denominator))
     return progress
 }
